Weapon overhaul, portal function
added stick, weapon switching, axe animation, portal scene switching

+++ EraPeril_IGB100Game/Assets/Scripts/NewScritps/WeaponData.cs
@@ -9,7 +9,7 @@ public class WeaponData : ScriptableObject{
     public new string name;
     public GameObject weaponModelObject;
     public int damage;
-    public int swingTime;
+    public float swingTime;
     public bool isThrowable;
     public AudioClip attackAudioClip;
